+++ NEWS.md
@@ -1,3 +1,3 @@
 # angstromATE 0.1.0
 
-* Initial CRAN submission.
+* Add Status XML reader